requirements.txt
pyautogui
PySide6
Pillow
pytesseract
pyperclip
python-dotenv
